refractor: move the utils functions and components out of the app route folder
